+++ src/v5/packet/publish.rs
@@ -12,7 +12,7 @@ use crate::{
         write::{Writable, wlen},
     },
     packet::{Packet, RxError, RxPacket, TxError, TxPacket},
-    types::{IdentifiedQoS, MqttString, QoS, TooLargeToEncode, TopicName, VarByteInt},
+    types::{IdentifiedQoS, MqttBinary, MqttString, QoS, TooLargeToEncode, TopicName, VarByteInt},
     v5::property::{
         AtMostOnceProperty, ContentType, CorrelationData, MessageExpiryInterval,
         PayloadFormatIndicator, Property, PropertyType, ResponseTopic, SubscriptionIdentifier,
@@ -254,6 +254,7 @@ impl<'p, const MAX_SUBSCRIPTION_IDENTIFIERS: usize>
     const EMPTY_TOPIC: MqttString<'static> = unsafe { MqttString::from_slice_unchecked("") };
 
     /// Creates a new packet with Quality of Service 0
+    #[allow(clippy::too_many_arguments)]
     pub fn new(
         dup: bool,
         retain: bool,
@@ -261,6 +262,8 @@ impl<'p, const MAX_SUBSCRIPTION_IDENTIFIERS: usize>
         message_expiry_interval: Option<MessageExpiryInterval>,
         topic: TopicReference<'p>,
         message: Bytes<'p>,
+        response_topic: Option<TopicName<'p>>,
+        correlation_data: Option<MqttBinary<'p>>,
     ) -> Result<Self, TooLargeToEncode> {
         let p = Self {
             dup,
@@ -269,8 +272,8 @@ impl<'p, const MAX_SUBSCRIPTION_IDENTIFIERS: usize>
             topic,
             payload_format_indicator: None,
             message_expiry_interval,
-            response_topic: None,
-            correlation_data: None,
+            response_topic: response_topic.map(Into::into),
+            correlation_data: correlation_data.map(Into::into),
             subscription_identifiers: Vec::new(),
             content_type: None,
             message,
@@ -352,6 +355,8 @@ mod unit {
                 TopicName::new_checked(MqttString::try_from("test/topic").unwrap()).unwrap(),
             ),
             Bytes::from("hello".as_bytes()),
+            None,
+            None,
         )
         .unwrap();
 
@@ -392,18 +397,23 @@ mod unit {
             Some(481123u32.into()),
             TopicReference::Alias(23408),
             Bytes::from("hello".as_bytes()),
+            Some(
+                TopicName::new_checked(MqttString::from_slice("uno, dos, tres, catorce").unwrap())
+                    .unwrap(),
+            ),
+            Some(MqttBinary::from_slice(&[0, 1, 2, 3, 4, 5, 6, 7]).unwrap()),
         )
         .unwrap();
 
         #[rustfmt::skip]
         encode!(packet, [
             0x3D,
-            0x12,
+            0x37,
             0x00, // Topic Name
             0x00, // Topic Name
             0x25, // Packet identifier
             0x98, // Packet identifier
-            0x08, // Property length
+            0x2D, // Property length
             0x02, // Message expiry interval
             0x00, //
             0x07, //
@@ -412,6 +422,15 @@ mod unit {
             0x23, // Topic alias
             0x5B, //
             0x70, // Topic alias
+
+            0x08, // Response Topic
+            0x00, 0x17,
+            b'u', b'n', b'o', b',', b' ', b'd', b'o', b's', b',', b' ', b't', b'r', b'e', b's', b',', b' ', b'c', b'a', b't', b'o', b'r', b'c', b'e', 
+            
+            0x09, // Correlation Data
+            0x00, 0x08,
+            0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07,
+
             b'h', // Payload
             b'e', //
             b'l', //
@@ -546,7 +565,7 @@ mod unit {
         assert_eq!(
             packet.response_topic,
             Some(ResponseTopic(
-                MqttString::try_from("response/topic").unwrap()
+                TopicName::new_checked(MqttString::try_from("response/topic").unwrap()).unwrap()
             ))
         );
         assert_eq!(