+++ tests/integration/creds.rs
@@ -11,6 +11,7 @@ const NO_CREDS_CONNECT_OPTIONS: ConnectOptions = ConnectOptions {
     clean_start: true,
     keep_alive: KeepAlive::Infinite,
     session_expiry_interval: SessionExpiryInterval::EndOnDisconnect,
+    request_response_information: false,
     user_name: None,
     password: None,
     will: None,